Update 12.02.2020
Replaced previous player sprite with another one, added melee attack pressing space which has an attached sphere in front of the player which has a radius which determines the enemies it hit with the attack. Added a new enemy "heavy bandit", made slight changes to tilemap parameters and spike obstacles. Rearanged some folders and scripts. Added a new library which has the new player and enemy model.